Make method documentation readable in IDEs
I.e. do not strip jsdoc in type definition files, but re-add the
documentation of methods to the type definition file to get inline
documentation in tools like Visual Studio Code.

This increases the size of the qooxdoo.d.ts file significantly, but
I think having proper method documentation in your IDE is worth it.

The change itself is rather small as it just prepends the jsdoc body
(properly split and prefixed with /**, * and **/) to the output of
each method, if there are jsdoc lines present.

This currently does not include the method params and return values,
though reconstructing wouldn't be a problem either. I've intentionally
left those out for now since they are sort of already documented in
the method prototype itself.