Fix class type detection and add getInstance to type definitions
The class type detection originally used was broken as it did not
properly check against the actual value of the property, but against
the type (StringValue or something). This resulted in the isSingleton,
isAbstract and isStatic properties not being set correctly on the
classes meta data field.

With this change it is possible to teach the TypeScriptWriter to add
a type definition for the getInstance method to singleton classes.