Adds --fix-jsdoc-params option to "qx lint"
This allows to change the order or @param name and {Type} to make it compatible for the generator 'name-first', `@param varname {Type}`) or with the correct JSDoc syntax ('type-first', `@param {Type} varname`).

Note that at the moment, the compiler and the APIViewer still rely on the (incorrect) syntax (see https://github.com/qooxdoo/qooxdoo-compiler/issues/446). This option makes it easy to switch between the syntaxes. Once the compiler accepts the JSDoc syntax, you can update your code by executing `qx lint --fix-jsdoc-params=type-first`.